Added New Approach to tile manager
After having moved around and cleaned up code once or twice I've ended up refactoring to this point. 

The GridManager keeps check of all buttons used for tiles, but also tracks the active grid item in question (so it knows which grid to draw tiles too, thus should allow multiple grids for foreground/background etc). It also handles mouse input for tiles and should handle the creation of new grids in time.

The grid object itself is mainly used to verify it's own coordinate values in both the array that keeps tiles and the world position of where to instantiate said tiles. Predominantly it handles floats to provide integer values from world space that can then be used on rounded world space values to create a "snap" effect and also provide the index value on the tile array.